requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
faster-whisper>=1.2.0
nemo-toolkit[asr]>=2.6.0
soundfile>=0.13.0
torch>=2.4.0
torchaudio>=2.4.0
transformers>=4.45.0
rapidfuzz>=3.10.0
numpy>=1.26.0
